Move FlowNetworkDataTypesInterface from event to backend package
