+++ core/src/backend/FlowDomainAgent.java
@@ -17,7 +17,6 @@
  */
 package backend;
 
-import event.FlowNetworkDataTypesInterface;
 import java.util.HashMap;
 import network.FlowNetwork;
 import org.apache.log4j.Logger;
