+++ core/src/backend/FlowNetworkDataTypesInterface.java
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  */
-package event;
+package backend;
 
 import network.Link;
 import network.Node;
