+++ core/src/input/EventLoaderNew.java
@@ -17,11 +17,11 @@
  */
 package input;
 
+import backend.FlowNetworkDataTypesInterface;
 import event.Event;
 import event.EventState;
 import event.NetworkComponent;
 import event.EventType;
-import event.FlowNetworkDataTypesInterface;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
