+++ core/src/input/FlowLoaderNew.java
@@ -5,7 +5,7 @@
  */
 package input;
 
-import event.FlowNetworkDataTypesInterface;
+import backend.FlowNetworkDataTypesInterface;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
